refactor: migrate @sanity/image-url to named export and main package types
