+++ src/sanity/lib/image.ts
@@ -1,11 +1,10 @@
-import createImageUrlBuilder from '@sanity/image-url'
-import { SanityImageSource } from "@sanity/image-url/lib/types/types";
+import { createImageUrlBuilder, SanityImageSource } from "@sanity/image-url";
 
-import { dataset, projectId } from '../env'
+import { dataset, projectId } from "../env";
 
 // https://www.sanity.io/docs/image-url
-const builder = createImageUrlBuilder({ projectId, dataset })
+const builder = createImageUrlBuilder({ projectId, dataset });
 
 export const urlFor = (source: SanityImageSource) => {
-  return builder.image(source)
-}
+  return builder.image(source);
+};
